+++ android/app/.idea/misc.xml
@@ -3,8 +3,9 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="app/src/main/res/layout/activity_main.xml" value="0.33" />
-        <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1875" />
+        <entry key="ruffle/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.3305" />
+        <entry key="ruffle/src/main/res/layout/activity_main.xml" value="0.33" />
+        <entry key="ruffle/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1875" />
       </map>
     </option>
   </component>
