+++ android/app/ruffle/src/main/res/layout/activity_main.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:ruffle="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".MainActivity">
+
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Ruffle"
+        android:textColor="#FF5722"
+        android:textSize="60sp"
+        ruffle:layout_constraintBottom_toTopOf="@+id/button"
+        ruffle:layout_constraintHorizontal_bias="0.5"
+        ruffle:layout_constraintLeft_toLeftOf="parent"
+        ruffle:layout_constraintRight_toRightOf="parent"
+        ruffle:layout_constraintTop_toTopOf="parent"
+        ruffle:layout_constraintVertical_bias="0.193" />
+
+    <Button
+        android:id="@+id/button"
+        android:layout_width="136dp"
+        android:layout_height="48dp"
+        android:text="Load an SWF"
+        ruffle:layout_constraintBottom_toBottomOf="parent"
+        ruffle:layout_constraintEnd_toEndOf="parent"
+        ruffle:layout_constraintHorizontal_bias="0.5"
+        ruffle:layout_constraintStart_toStartOf="parent"
+        ruffle:layout_constraintTop_toBottomOf="@+id/textView" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
